+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Soft delete games when they expire instead of actually deleting them.
+
 ## [v7.6.1](https://github.com/lexicalunit/spellbot/releases/tag/v7.6.1) - 2021-11-25
 
 ### Fixed